+++ runtime/runtime/types.h
@@ -2,18 +2,16 @@
 #ifndef BASALT_RUNTIME_TYPES_H
 #define BASALT_RUNTIME_TYPES_H
 
-#include "Input.h"
-
-#include "shared/Size2D.h"
 #include "shared/Types.h"
 
 namespace basalt {
 
-struct UpdateContext final {
-  f64 deltaTime {};
-  Size2Du16 windowSize {Size2Du16::dont_care()};
-  Input input {};
+enum class MouseCursor : u8 {
+  Arrow, TextInput, ResizeAll, ResizeNS, ResizeEW, ResizeNESW, ResizeNWSE, Hand
+, NotAllowed
 };
+
+constexpr uSize MOUSE_CURSOR_COUNT = 9u;
 
 } // namespace basalt